Add files via upload
The index.tsx file is used to view the additional details for the course, the page.tsx in courses is to show the different sections for information regarding resources, course requirements for major and minor. The courseCode directory containing the page.tsx file will list the credits, hours, and prerequisites for the course.